MT#55283 unify types used for sequence numbers
We use int or unsigned int everywhere, except in a few leftover places.
Unify to (unsigned) int.

(cherry picked from commit f60ee917c9d08717299adabf82de7ddeede259ba)

+++ daemon/media_socket.c
@@ -2694,10 +2694,10 @@ static int stream_packet(struct packet_handler_ctx *phc) {
 		atomic64_inc_na(&phc->mp.ssrc_in->stats->packets);
 		atomic64_add_na(&phc->mp.ssrc_in->stats->bytes, phc->s.len);
 		// no real sequencing, so this is rudimentary
-		uint64_t old_seq = atomic_get_na(&phc->mp.ssrc_in->stats->ext_seq);
-		uint64_t new_seq = ntohs(phc->mp.rtp->seq_num) | (old_seq & 0xffff0000UL);
+		unsigned int old_seq = atomic_get_na(&phc->mp.ssrc_in->stats->ext_seq);
+		unsigned int new_seq = ntohs(phc->mp.rtp->seq_num) | (old_seq & 0xffff0000UL);
 		// XXX combine this with similar code elsewhere
-		long seq_diff = new_seq - old_seq;
+		int seq_diff = new_seq - old_seq;
 		while (seq_diff < -60000) {
 			new_seq += 0x10000;
 			seq_diff += 0x10000;
